+++ src/components/common/guidePage/GuidePage.type.ts
@@ -1,0 +1,10 @@
+export interface GettingStartedType {
+  className: string
+  showHelpCard: boolean
+  hideGettingStartedCard: (count?: string) => void
+  loginCount: number
+}
+
+export interface GuidedPageType {
+  openDevtronAppCreateModel: (event) => void
+}
