+++ src/components/onboardingGuide/OnboardingGuide.type.ts
@@ -1,0 +1,20 @@
+export interface CommonGuideType{
+  loginCount: number
+  title: string
+  subtitle: string
+  onClickCloseButton: () => void
+  isGettingStartedClicked: boolean
+}
+
+export interface DeployManageGuide{
+  isGettingStartedClicked: boolean
+  loginCount: number
+}
+
+export interface OnboardingGuideProps {
+  loginCount: number
+  isSuperAdmin: boolean
+  serverMode: string
+  onClickedDeployManageCardClicked: () => void
+  isGettingStartedClicked: boolean
+}
