+++ src/components/onboardingGuide/onboardingGuide.scss
@@ -23,7 +23,7 @@
           text-decoration: none;
         }
         .onboarding-card:hover{
-          border: 1px solid var(--B500);
+          box-shadow: 0 8px 12px 0 rgb(30 35 96 / 10%), 0 1px 4px 0 rgb(0 0 0 / 10%);
           border-radius: 400;
         }
      }
